Fix: bad function usage in README for UltiSnips
This commit fixes an issue with using vsnip#anonymous in the snippet expand function for Ultisnip
The main issue, is that when using vsnip#anonymous instead of UltiSnip#Anon when calling the cmp.mapping.confirm map on a non-snippet option would "crash" cmp, ceasing its normal functionality. Using UltiSnps#Anon fixes this issue.