Make the pum window property respect future changes of &pumblend
PR #310 added a support for the pumblend property for the custom
floating completion pop-up menu, but the winblend option is set
globally which doesn't respect future changes (e.g., `set pumblend=..`).

Such window options for the pop-up window that might be changed later
would need to be set when the floating window is being opened.