Check for complete pattern matches later in the target text
This allows us to find the best match when trying to match "so"
against "schemas/source_files.ex". Without this, bescause the 's' in
"so" matches against "schemas", we never find the complete match against
"source_files" later in the target. Look for a later total match and add
it to the list if it is at least as good as the computed partial match
so far. If we also have a partial-match continuation, prefer the longer
of the 2 choices.