fix: remove redundant UIEnter subscription
The UIEnter event is not needed for setting up the
highlight groups in cmp.lua.

This caused a bug displaying ANSI escape sequences in
neovim in tmux, kitty, and alacritty.

I found this bug after 2 days of debugging, searching for
the cause in the wrong place - zsh, tmux, kitty, then
testing on alacritty and xterm, and finally realizing
that the issue was in the Neovim configuration itself.

If this does not in fact fix the issue, even though I
tested it thoroughly, it is still a redundant call
in my opinion, especially since `ColorScheme` event
is immediately triggered after the subscription.